refactor: mode property move to types
